refactor: Renamed semantic to version and included semantic as part of the overall module.
