+++ src/version/mod.rs
@@ -1,0 +1,10 @@
+mod level;
+mod pre_release;
+mod semantic;
+pub(crate) mod test_utils;
+mod version_tag;
+
+pub use level::Level;
+pub(crate) use pre_release::{PreRelease, PreReleaseType};
+pub use semantic::Semantic;
+pub use version_tag::VersionTag;
