+++ src/version/version_tag.rs
@@ -205,7 +205,7 @@ mod tests {
 
     use rstest::rstest;
 
-    use crate::semantic::{test_utils::gen_current_version, PreRelease};
+    use crate::version::{test_utils::gen_current_version, PreRelease};
 
     use super::*;
